requirements.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
cached-property==1.5.2
celery==5.1.2
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.5
djangorestframework==3.12.4
dnspython==1.16.0
importlib-metadata==4.6.1
kombu==5.1.0
prompt-toolkit==3.0.19
psycopg2==2.9.1
pykerberos==1.2.1
pymongo==3.11.4
python-decouple==3.4
python-snappy==0.6.0
pytz==2021.1
redis==3.5.3
redis-server==6.0.9
six==1.16.0
sqlparse==0.4.1
supertools==1.0.1
typing-extensions==3.10.0.0
vine==5.0.0
wcwidth==0.2.5
zipp==3.5.0
